+++ wayland-protocols-plasma/src/lib.rs
@@ -42,6 +42,15 @@ pub mod dpms {
     );
 }
 
+pub mod external_brightness {
+    pub mod v1 {
+        wayland_protocol!(
+            "./plasma-wayland-protocols/src/protocols/kde-external-brightness-v1.xml",
+            []
+        );
+    }
+}
+
 pub mod fake_input {
     wayland_protocol!(
         "./plasma-wayland-protocols/src/protocols/fake-input.xml",
@@ -121,7 +130,7 @@ pub mod plasma_shell {
 
 pub mod plasma_virtual_desktop {
     wayland_protocol!(
-        "./plasma-wayland-protocols/src/protocols/plasma-virtual-desktop.xml",
+        "./plasma-wayland-protocols/src/protocols/org-kde-plasma-virtual-desktop.xml",
         []
     );
 }
@@ -144,6 +153,15 @@ pub mod screencast {
     pub mod v1 {
         wayland_protocol!(
             "./plasma-wayland-protocols/src/protocols/zkde-screencast-unstable-v1.xml",
+            []
+        );
+    }
+}
+
+pub mod screen_edge {
+    pub mod v1 {
+        wayland_protocol!(
+            "./plasma-wayland-protocols/src/protocols/kde-screen-edge-v1.xml",
             []
         );
     }